購入処理・トランザクションの記述

+++ app/Http/Controllers/OrderController.php
@@ -4,11 +4,13 @@ namespace App\Http\Controllers;
 
 use App\Models\Cart;
 use App\Models\Order;
+use App\Models\OrderDetail;
 use App\Models\User;
 use App\Models\UserInfo;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
+use Livewire\Features\SupportConsoleCommands\Commands\CopyCommand;
 
 class OrderController extends Controller
 {
@@ -28,22 +30,44 @@ class OrderController extends Controller
 
     public function store()
     {
+        $userInfo = UserInfo::where('user_id', Auth::id())->first();
+        if (!$userInfo) {
+            return redirect()->route('order.confirm');
+        }
+
         try {
 
             DB::transaction(function () {
 
+                $cartItems = Cart::with('item')->where('user_id', Auth::id())->get();
+
+                $order = Order::create([
+
+                    'user_id' => Auth::id()
+                ]);
+
+                foreach ($cartItems as $cartItem) {
+
+                    OrderDetail::create([
+
+                        'order_id' => $order->id,
+                        'item_id' => $cartItem->item->id,
+                        'item_num' => $cartItem->item_num
+                    ]);
+                }
+
                 Cart::where('user_id', Auth::id())->delete();
 
-                Order::create([
-                    'login_id' => login_id,
-                    'password' => password,
-                ]);
+                $orders = Order::with('order_details.item')->where('user_id', Auth::id())->where('id', $order->id)->get();
+                
+                return redirect()->route('order.thank_you', compact('orders'));
             }, 5);
         } catch (\Exception $e) {
-            Log::error('購入時のエラー');
+
+            Log::error('購入処理中のエラー' . $e);
+            return redirect()->route('errors.error');
         }
 
-        return redirect()->route('order.thank_you');
     }
 
     public function thankYou()
